Add files via upload
'Design of the project.pdf', 'Implementation Details.pdf', 'Project Synopsis' and 'References.pdf' are some files, which stores some information of the mini-project. main.c is the program. main.exe is the file which opens in powershell, and is the face of the whole project, where changes and operations are reflected. record.dat is the file which contains all the data used by our code.